Refactor stats tracking handler package and update save/load logic
Renamed StatsTrackingButtonHandler from the 'tigl' package to 'statistics' for clarity. Updated all references accordingly. Adjusted GameSaveService and GameLoadService to encode spaces as '----' in stats tracked user names instead of user IDs, ensuring consistent serialization and deserialization.
