+++ src/main/java/ti4/buttons/handlers/statistics/StatsTrackingButtonHandler.java
@@ -1,4 +1,4 @@
-package ti4.buttons.handlers.tigl;
+package ti4.buttons.handlers.statistics;
 
 import lombok.experimental.UtilityClass;
 import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
