+++ src/main/java/ti4/commands/game/Replace.java
@@ -19,7 +19,7 @@ import net.dv8tion.jda.api.interactions.commands.OptionType;
 import net.dv8tion.jda.api.interactions.commands.build.OptionData;
 import org.apache.commons.lang3.function.Consumers;
 import ti4.buttons.Buttons;
-import ti4.buttons.handlers.tigl.StatsTrackingButtonHandler;
+import ti4.buttons.handlers.statistics.StatsTrackingButtonHandler;
 import ti4.commands.CommandHelper;
 import ti4.commands.GameStateSubcommand;
 import ti4.helpers.Constants;
