+++ src/main/java/ti4/map/persistence/GameLoadService.java
@@ -933,9 +933,9 @@ class GameLoadService {
                 case Constants.COLOR -> player.setColor(tokenizer.nextToken());
                 case Constants.DECAL_SET -> player.setDecalSet(tokenizer.nextToken());
                 case Constants.STATS_ANCHOR_LOCATION -> player.setPlayerStatsAnchorPosition(tokenizer.nextToken());
-                case Constants.STATS_TRACKED_USER_ID ->
-                    player.setStatsTrackedUserID(tokenizer.nextToken().replace("----", " "));
-                case Constants.STATS_TRACKED_USER_NAME -> player.setStatsTrackedUserName(tokenizer.nextToken());
+                case Constants.STATS_TRACKED_USER_ID -> player.setStatsTrackedUserID(tokenizer.nextToken());
+                case Constants.STATS_TRACKED_USER_NAME ->
+                    player.setStatsTrackedUserName(tokenizer.nextToken().replace("----", " "));
                 case Constants.HS_TILE_POSITION -> player.setHomeSystemPosition(tokenizer.nextToken());
                 case Constants.ALLIANCE_MEMBERS -> player.setAllianceMembers(tokenizer.nextToken());
                 case Constants.ROLE_FOR_COMMUNITY -> player.setRoleIDForCommunity(tokenizer.nextToken());
