+++ src/main/java/ti4/map/persistence/GameSaveService.java
@@ -749,11 +749,11 @@ class GameSaveService {
             writer.write(Constants.STATS_ANCHOR_LOCATION + " " + player.getPlayerStatsAnchorPosition());
             writer.write(System.lineSeparator());
 
-            writer.write(Constants.STATS_TRACKED_USER_ID + " "
-                    + player.getStatsTrackedUserID().replace(" ", "----"));
+            writer.write(Constants.STATS_TRACKED_USER_ID + " " + player.getStatsTrackedUserID());
             writer.write(System.lineSeparator());
 
-            writer.write(Constants.STATS_TRACKED_USER_NAME + " " + player.getStatsTrackedUserName());
+            writer.write(Constants.STATS_TRACKED_USER_NAME + " "
+                    + player.getStatsTrackedUserName().replace(" ", "----"));
             writer.write(System.lineSeparator());
 
             writer.write(Constants.HS_TILE_POSITION + " " + player.getHomeSystemPosition());
